Move reusable health check logic to web framework project.
